scope materialization runs state per editor tab
