+++ frontend/src/scenes/data-warehouse/editor/output-pane-tabs/infoTabLogic.ts
@@ -1,10 +1,18 @@
-import { connect, kea, key, path, props, selectors } from 'kea'
+import { actions, connect, kea, key, listeners, path, props, reducers, selectors } from 'kea'
+import { loaders } from 'kea-loaders'
+import { subscriptions } from 'kea-subscriptions'
 
+import api, { PaginatedResponse } from 'lib/api'
 import { databaseTableListLogic } from 'scenes/data-management/database/databaseTableListLogic'
 import { dataWarehouseViewsLogic } from 'scenes/data-warehouse/saved_queries/dataWarehouseViewsLogic'
 
+import { DataModelingJob } from '~/types'
+
 import { sqlEditorLogic } from '../sqlEditorLogic'
 import type { infoTabLogicType } from './infoTabLogicType'
+
+const REFRESH_INTERVAL = 10000
+const DEFAULT_JOBS_PAGE_SIZE = 10
 
 export interface InfoTableRow {
     name: string
@@ -25,13 +33,66 @@ export const infoTabLogic = kea<infoTabLogicType>([
     connect((props: InfoTabLogicProps) => ({
         values: [
             sqlEditorLogic({ tabId: props.tabId }),
-            ['metadata'],
+            ['metadata', 'editingView'],
             databaseTableListLogic,
             ['posthogTablesMap', 'dataWarehouseTablesMap'],
             dataWarehouseViewsLogic,
             ['dataWarehouseSavedQueryMap'],
         ],
     })),
+    actions({
+        setStartingMaterialization: (starting: boolean) => ({ starting }),
+    }),
+    loaders(({ values }) => ({
+        dataModelingJobs: [
+            null as PaginatedResponse<DataModelingJob> | null,
+            {
+                loadDataModelingJobs: async (savedQueryId: string) => {
+                    return await api.dataWarehouseSavedQueries.dataWarehouseDataModelingJobs.list(
+                        savedQueryId,
+                        values.dataModelingJobs?.results.length
+                            ? Math.max(values.dataModelingJobs.results.length, DEFAULT_JOBS_PAGE_SIZE)
+                            : DEFAULT_JOBS_PAGE_SIZE,
+                        0
+                    )
+                },
+                loadOlderDataModelingJobs: async () => {
+                    const nextUrl = values.dataModelingJobs?.next
+
+                    if (!nextUrl) {
+                        return values.dataModelingJobs
+                    }
+
+                    const res = await api.get<PaginatedResponse<DataModelingJob>>(nextUrl)
+                    res.results = [...(values.dataModelingJobs?.results ?? []), ...res.results]
+
+                    return res
+                },
+                resetDataModelingJobs: () => null,
+            },
+        ],
+    })),
+    reducers({
+        startingMaterialization: [
+            false,
+            {
+                setStartingMaterialization: (_, { starting }: { starting: boolean }) => starting,
+                loadDataModelingJobsSuccess: (
+                    state: boolean,
+                    { dataModelingJobs }: { dataModelingJobs: PaginatedResponse<DataModelingJob> | null }
+                ) => {
+                    const currentJobStatus = dataModelingJobs?.results?.[0]?.status
+                    if (
+                        currentJobStatus &&
+                        ['Running', 'Completed', 'Failed', 'Cancelled'].includes(currentJobStatus)
+                    ) {
+                        return false
+                    }
+                    return state
+                },
+            },
+        ],
+    }),
     selectors({
         sourceTableItems: [
             (s) => [s.metadata, s.dataWarehouseSavedQueryMap],
@@ -62,5 +123,27 @@ export const infoTabLogic = kea<infoTabLogicType>([
                 )
             },
         ],
+        hasMoreJobsToLoad: [(s) => [s.dataModelingJobs], (dataModelingJobs) => !!dataModelingJobs?.next],
     }),
+    listeners(({ actions, cache }) => ({
+        loadDataModelingJobsSuccess: ({ payload }) => {
+            cache.disposables.add(() => {
+                const timeoutId = setTimeout(() => {
+                    if (payload) {
+                        actions.loadDataModelingJobs(payload)
+                    }
+                }, REFRESH_INTERVAL)
+                return () => clearTimeout(timeoutId)
+            }, 'dataModelingJobsRefreshTimeout')
+        },
+    })),
+    subscriptions(({ actions, values }) => ({
+        editingView: (editingView) => {
+            if (editingView) {
+                if (values.dataModelingJobs === null) {
+                    actions.loadDataModelingJobs(editingView.id)
+                }
+            }
+        },
+    })),
 ])
